update to allow most up-to-date scipy, numpy, pandas, and anndata

+++ tests/integration/conftest.py
@@ -245,8 +245,7 @@ def raw_reference_h5ad_fixture(
         a_data = anndata.AnnData(
             X=raw_reference_cell_x_gene,
             obs=obs,
-            var=var,
-            dtype=raw_reference_cell_x_gene.dtype)
+            var=var)
 
     h5ad_path = pathlib.Path(
         mkstemp_clean(
@@ -278,8 +277,7 @@ def noisier_raw_reference_h5ad_fixture(
         a_data = anndata.AnnData(
             X=noisier_raw_reference_cell_x_gene,
             obs=obs,
-            var=var,
-            dtype=noisier_raw_reference_cell_x_gene.dtype)
+            var=var)
 
     h5ad_path = pathlib.Path(
         mkstemp_clean(
@@ -344,8 +342,7 @@ def raw_query_h5ad_fixture(
         a_data = anndata.AnnData(
             X=raw_query_cell_x_gene_fixture,
             var=var,
-            uns={'AIBS_CDM_gene_mapping': {'a': 'b', 'c': 'd'}},
-            dtype=raw_query_cell_x_gene_fixture.dtype)
+            uns={'AIBS_CDM_gene_mapping': {'a': 'b', 'c': 'd'}})
 
     h5ad_path = pathlib.Path(
         mkstemp_clean(
@@ -577,8 +574,7 @@ def noisy_raw_reference_h5ad_fixture(
         a_data = anndata.AnnData(
             X=scipy.sparse.csr_matrix(data),
             obs=obs,
-            var=var,
-            dtype=int)
+            var=var)
 
     h5ad_path = pathlib.Path(
         mkstemp_clean(
@@ -642,8 +638,7 @@ def noisy_raw_query_h5ad_fixture(
         a_data = anndata.AnnData(
             X=data,
             var=var,
-            uns={'AIBS_CDM_gene_mapping': {'a': 'b', 'c': 'd'}},
-            dtype=int)
+            uns={'AIBS_CDM_gene_mapping': {'a': 'b', 'c': 'd'}})
 
     h5ad_path = pathlib.Path(
         mkstemp_clean(
