+++ tests/utils/test_csc_to_csr.py
@@ -62,7 +62,7 @@ def csc_fixture(
         warnings.simplefilter('ignore')
         a = anndata.AnnData(
                 X=scipy_sparse.csc_matrix(x_array_fixture),
-                dtype=x_array_fixture.dtype)
+        )
     a.write_h5ad(h5ad_path)
     with h5py.File(h5ad_path, 'r') as src:
         attrs = dict(src['X'].attrs)
